Move relevant binary crate functions out into proper binary crate
